Change default value of 3rd parameter to 2nd parameter (#140)
IMO it is a bad idea to have the 3rd parameter default to ", ", I propose it to default to the 2nd parameter.
Using `[[TO=lastname| & ]]` with Mr. Smith and Mr. Jones currently results to `Smith, Jones` instead of my expectation: `Smith & Jones`.

As a workaround I have to use `[[TO=lastname| & | & ]]`.
IMO it should be fixed and published soon, as this breaks (broke) existing templates.